+++ internal/publish/connect_cloud/preflight_checks.go
@@ -1,0 +1,8 @@
+package connect_cloud
+
+// Copyright (C) 2025 by Posit Software, PBC.
+
+func (c *ServerPublisher) PreFlightChecks() error {
+	// No-op
+	return nil
+}